services/discovery/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0.1 